Follow the principle of maximal munch
Previously, scan_token() in class `Scanner` tokenized keyword `or` like
operators of length 2, such as `//` or `!=`. As a result, identifiers
starting with "or" -- for example "orchid" as provided by Robert Nystrom
in the book -- are treated as the conditional operator, leading to
unintended consequences.

It was a silly oversight on my part to tokenize `or` with this method
because lexemes must be tokenized as the longest or maximal token, hence
the name of the principle of maximal munch. Nystrom included it in the
book specifically as an example of a method to *avoid*. Keyword `and`
was not tokenized this way prior either. The lexeme was searched for in
a table of keywords. If found, the lexeme was tokenized as the
respective keyword, as `or` is now. Otherwise, it was tokenized as an
identifier to, say, a variable or a function.